Use `expect( console ).toHaveWarned()` instead of mocking `@wordpress/deprecated`, move new tests into one file
